Refactor files.mk to avoid manually specifying all files
This refactors opengapps-files.mk to avoid specifying every file, replacing
it instead with two functions that search and find all the necessary files
to copy over.

Also remove AndroidProducts.mk since that file is not supposed to be used.
It's a remnant of the original code generation solution.

Related to #9