Configure network location provider overlay for all types
The network location provider must be configured in the XML for all
device types, otherwise issues occur on 6.0+ with permissions, and
on all devices with the network location provider.

I'm not sure how this will go in CyanogenMod builds since there
vendor/cm already defines a similar overlay...It should go OK thought.

This should fix #12 and #35