Always use the highest compatible API level.
Some apps (at least Play Services) will not work if they have multiple api levels and the one installed is not the highest API the device can take.
This cannot be detected from aapt because the min and target sdk in the manifest are the same.
It is therefore better to favor the correct API level instead a more recent package which might not work.