Start using the path "vendor/opengapps/build"
The "aosp_build" project was previously installed under the path
"vendor/google/build".  This was done to take advantage of a
google-specific include statement in build/core/main.mk:

  # Include the google-specific config
  -include vendor/google/build/config.mk

However, this caused problems because it meant that the opengapps
build files were always processed, even when building for a device
that should not have opengapps added to it; it also caused some
generic makefile targets like "make clean" to break (unless
GAPPS_VARIANT was set).

Start using the path "vendor/opengapps/build" for "aosp_build" and
avoid issues related to unwanted processing of "config.mk".