Futher restrict invalid module error
Only emit the error if the module is in the set-difference
GAPPS_PRODUCT_PACKAGES - GAPPS_EXCLUDED_PACKAGES.  Also update
README.md to suggest using GAPPS_PRODUCT_PACKAGES rather than
PRODUCT_PACKAGES to add modules; if you happen to append an invalid
module to PRODUCT_PACKAGES, then the error won't be emitted.  Note
that PRODUCT_PACKAGES is not defined when the build system does the
initial processing of Android.mk files.