Add module for AndroidMigratePrebuilt
AndroidMigratePrebuilt, also known as Data Transfer Tool, is launched
by SetupWizard.  If it is not pre-installed, then it may be downloaded
and installed by SetupWizard.  In that case, the downloaded version
does not receive the required permissions and will crash when run.
The pre-installed version is granted the necessary permissions because
of an entry in the opengapps file privapp-permissions-google.xml:

  https://github.com/GraphiteSoftware/all/blob/master/etc/permissions/privapp-permissions-google.xml