find_apk: Replace `rev` with longest substring removal for DPI parsing.
rev is not available as path tool in the Android 10 build system:
https://android.googlesource.com/platform/build/+/master/Changes.md#PATH_Tools

While this will only generate warnings for the time being, extracting
the first and last number from a dpi sequence a-b-c can be implemented
natively in bash without requiring a subprocess. Note that _longest_
prefix/postfix match is used, to strip off anything before/after the
dash.

Signed-off-by: MarijnS95 <marijns95@gmail.com>