overlay: Move webview overlay from device/vendor to GoogleWebView RRO.
Replicate GoogleWebViewOverlay.apk that contains the webviewprovider
overlay and install it to the system partition, just like all other
Gapps.
This overlay would otherwise end up in the device-specific RRO (or built
directly into the framework when the device doesn't enforce an RRO for
framework-res) where it doesn't belong.

As far as I'm aware RRO packages have only recently been added in Oreo
(SDK 26).  A fallback for Nougat has been kept that adds the moved file
to the device overlay.

Signed-off-by: MarijnS95 <marijns95@gmail.com>