Include trichromelibrary with Chrome/WebView on SDK >= 29.
Chrome and the WebViewChrome library have all their shared code moved to
TrichromeLibrary. Add a prebuilt module for this library, which is
required to run either apk.

Without this, Chrome and anything using WebView will crash (unless you
select a different WebView or install it fresh from the Play Store).

Signed-off-by: MarijnS95 <marijns95@gmail.com>